Багфиксы, удалено немного ненужного говна.

+++ code/game/objects/items/stacks/stack.dm
@@ -138,14 +138,17 @@
 	return
 
 /obj/item/stack/proc/use(var/amount)
+	if(src.amount < amount)
+		return 0
+
 	src.amount-=amount
-	if (src.amount<=0)
+	if(src.amount<=0)
 		var/oldsrc = src
 		src = null //dont kill proc after del()
 		if(usr)
 			usr.before_take_item(oldsrc)
 		del(oldsrc)
-	return
+	return 1
 
 /obj/item/stack/proc/add_to_stacks(mob/usr as mob)
 	var/obj/item/stack/oldsrc = src
